+++ src/graphql/documents/fragments.ts
